@@ -1,9 +1,9 @@
 apply plugin: 'com.android.library'
-apply plugin: 'com.novoda.bintray-release'
+//apply plugin: 'com.novoda.bintray-release'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.1"
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         minSdkVersion 10
@@ -21,15 +21,18 @@ android {
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     testCompile 'junit:junit:4.12'
-    compile 'com.android.support:recyclerview-v7:23.4.0'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    api 'androidx.recyclerview:recyclerview:1.1.0'
 }
 
-publish {
-    userOrg = 'hongyangandroid'
-    groupId = 'com.zhy'
-    artifactId = 'base-rvadapter'
-    publishVersion = '3.0.3'
-    desc = 'a easy adapter for recyclerview'
-    website = 'https://github.com/hongyangAndroid/baseAdapter'
-    licences = ['Apache-2.0']
-}
+apply from: './maven_push.gradle'
+
+//publish {
+//    userOrg = 'hongyangandroid'
+//    groupId = 'com.zhy'
+//    artifactId = 'base-rvadapter'
+//    publishVersion = '3.0.3'
+//    desc = 'a easy adapter for recyclerview'
+//    website = 'https://github.com/hongyangAndroid/baseAdapter'
+//    licences = ['Apache-2.0']
+//}
