@@ -1,8 +1,9 @@
 package com.zhy.sample;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.widget.ListView;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.zhy.adapter.abslistview.CommonAdapter;
 import com.zhy.sample.adapter.lv.ChatAdapter;
